refactor: extract SplitScrollCoordinator into its own file
Preparatory cleanup before DocumentSurfaceViewModel extraction.
Purely structural — no behavior changes.
